Send MQTT message also when no plate is detected
This is required for my HASS automation - right now, HASS only receives a message only when a plate has been detected.
There are situations where 'car' object has been found, but Plate Recognizer couldn't get a plate (e.g. the car is sideways and it can't see the plate). I still want to receive a notification in those situations, so an empty message is sent with only event and review id.